use react form, and updated the type
