+++ client/src/api/axios.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
 
 export default axios.create({
-  baseURL: 'http://localhost:8080/api/',
+  baseURL: 'http://localhost:8081/api/',
 });
