+++ client/src/pages/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Button } from '@/components/ui/button';
-import CreateNoticeForm from '@/components/CreateNoticeForm';
+import CreateNoticeForm from '@/components/NoticeForm';
 
 const Home = () => {
   return (
