+++ client/src/types/type.ts
@@ -1,0 +1,17 @@
+export type RelatedLink = {
+    title: string;
+    url: string;
+  };
+  
+  export type NoticeFormData = {
+    title: string;
+    description: string;
+    pdfLink: string;
+    postedDate: string;
+    expiryDate: string;
+    author: string;
+    tags: string;
+    status: string;
+    relatedLinks: RelatedLink[];
+  };
+
