+++ server/src/controllers/noticeController.ts
