+++ server/src/index.ts
@@ -10,7 +10,7 @@ import noticeRoute from './routes/noticeRoutes';
 dotenv.config();
 
 const app = express();
-const port = process.env.PORT || 8080;
+const port = process.env.PORT || 8081;
 
 app.use(express.json());
 app.use(cors());
