+++ server/src/routes/noticeRoutes.ts
